Use identifier in time-series data

+++ homeserver-common/src/main/java/org/juhanir/domain/sensordata/dto/outgoing/TemperatureStatusResponse.java
@@ -17,8 +17,8 @@ public record TemperatureStatusResponse(
     public static TemperatureStatusResponse fromTemperatureStatus(TemperatureStatus tempStatus) {
         return new TemperatureStatusResponse(
                 tempStatus.getComponentId(),
-                tempStatus.getValueCelsius().doubleValue(),
-                tempStatus.getValueFahrenheit().doubleValue(),
+                tempStatus.getValueCelsius(),
+                tempStatus.getValueFahrenheit(),
                 tempStatus.getMeasurementTime(),
                 tempStatus.getDevice().getIdentifier()
         );
